+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Set visibility to "Followers" when self-replying to followers-only post.
 - Enable username autocomplete on login form.
 - Don't allow to edit posts with text/html content type.
+- Expect status code 200 when deleting post.
 
 ### Fixed