+++ CHANGELOG.md
@@ -13,7 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 
-- Updated translations ([#51](https://codeberg.org/silverpill/mitra-web/pulls/51)).
+- Updated translations ([#51](https://codeberg.org/silverpill/mitra-web/pulls/51), [#52](https://codeberg.org/silverpill/mitra-web/pulls/52)).
 - Add tooltip to disabled visibility selector.
 - Set visibility to "Followers" when self-replying to followers-only post.
 - Enable username autocomplete on login form.