+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Move "Import follows" and "Move followers" buttons out of "Experiments" section.
 - Prepare import and migration pages for translation.
 - Try to fetch profile before moving followers.
+- Updated translations ([#54](https://codeberg.org/silverpill/mitra-web/pulls/54)).
 
 ## [3.10.0] - 2024-11-20