+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.11.0] - 2024-12-05
+
 ### Changed
 
 - Move "Import follows" and "Move followers" buttons out of "Experiments" section.